Update endpoints to use full relevancy confidence and model name as well
