+++ src/api/endpoints/review/next/dto.py
@@ -2,13 +2,14 @@ from typing import Optional
 
 from pydantic import BaseModel, Field
 
-from src.api.endpoints.annotate.dtos.agency.response import GetNextURLForAgencyAgencyInfo
+from src.api.endpoints.annotate.agency.get.dto import GetNextURLForAgencyAgencyInfo
+from src.api.endpoints.annotate.relevance.get.dto import RelevanceAnnotationResponseInfo
 from src.core.enums import RecordType, SuggestedStatus
 from src.core.tasks.url.operators.url_html.scraper.parser.dtos.response_html import ResponseHTMLInfo
 
 
 class FinalReviewAnnotationRelevantInfo(BaseModel):
-    auto: Optional[bool] = Field(title="Whether the auto-labeler has marked the URL as relevant")
+    auto: Optional[RelevanceAnnotationResponseInfo] = Field(title="Whether the auto-labeler has marked the URL as relevant")
     user: Optional[SuggestedStatus] = Field(
         title="The status marked by a user, if any",
     )
