+++ src/api/endpoints/review/routes.py
@@ -1,15 +1,13 @@
-from typing import Optional
-
 from fastapi import APIRouter, Depends, Query
 
 from src.api.dependencies import get_async_core
-from src.api.endpoints.review.dtos.approve import FinalReviewApprovalInfo
-from src.api.endpoints.review.dtos.get import GetNextURLForFinalReviewOuterResponse
-from src.api.endpoints.review.dtos.reject import FinalReviewRejectionInfo
+from src.api.endpoints.review.approve.dto import FinalReviewApprovalInfo
+from src.api.endpoints.review.next.dto import GetNextURLForFinalReviewOuterResponse
+from src.api.endpoints.review.reject.dto import FinalReviewRejectionInfo
 from src.core.core import AsyncCore
-from src.security.manager import require_permission
 from src.security.dtos.access_info import AccessInfo
 from src.security.enums import Permissions
+from src.security.manager import require_permission
 
 review_router = APIRouter(
     prefix="/review",
@@ -19,14 +17,17 @@ review_router = APIRouter(
 
 requires_final_review_permission = require_permission(Permissions.SOURCE_COLLECTOR_FINAL_REVIEW)
 
+batch_id_query = Query(
+    description="The batch id of the next URL to get. "
+                "If not specified, defaults to first qualifying URL",
+    default=None
+)
+
 @review_router.get("/next-source")
 async def get_next_source(
     core: AsyncCore = Depends(get_async_core),
     access_info: AccessInfo = Depends(requires_final_review_permission),
-    batch_id: Optional[int] = Query(
-        description="The batch id of the next URL to get. "
-                    "If not specified, defaults to first qualifying URL",
-        default=None),
+    batch_id: int | None = batch_id_query,
 ) -> GetNextURLForFinalReviewOuterResponse:
     return await core.get_next_source_for_review(batch_id=batch_id)
 
@@ -35,10 +36,7 @@ async def approve_source(
     core: AsyncCore = Depends(get_async_core),
     access_info: AccessInfo = Depends(requires_final_review_permission),
     approval_info: FinalReviewApprovalInfo = FinalReviewApprovalInfo,
-    batch_id: Optional[int] = Query(
-        description="The batch id of the next URL to get. "
-                    "If not specified, defaults to first qualifying URL",
-        default=None),
+    batch_id: int | None = batch_id_query,
 ) -> GetNextURLForFinalReviewOuterResponse:
     await core.approve_url(
         approval_info,
@@ -51,10 +49,7 @@ async def reject_source(
     core: AsyncCore = Depends(get_async_core),
     access_info: AccessInfo = Depends(requires_final_review_permission),
     review_info: FinalReviewRejectionInfo = FinalReviewRejectionInfo,
-    batch_id: Optional[int] = Query(
-        description="The batch id of the next URL to get. "
-                    "If not specified, defaults to first qualifying URL",
-        default=None),
+    batch_id: int | None = batch_id_query,
 ) -> GetNextURLForFinalReviewOuterResponse:
     await core.reject_url(
         url_id=review_info.url_id,
