+++ tests/automated/integration/api/review/conftest.py
@@ -1,6 +1,6 @@
 import pytest_asyncio
 
-from src.api.endpoints.annotate.dtos.agency.post import URLAgencyAnnotationPostInfo
+from src.api.endpoints.annotate.agency.post.dto import URLAgencyAnnotationPostInfo
 from src.collectors.enums import URLStatus
 from src.core.enums import SuggestedStatus, RecordType
 from tests.helpers.batch_creation_parameters.annotation_info import AnnotationInfo
