+++ tests/automated/integration/api/review/rejection/helpers.py
@@ -1,6 +1,6 @@
-from src.api.endpoints.review.dtos.get import GetNextURLForFinalReviewOuterResponse
-from src.api.endpoints.review.dtos.reject import FinalReviewRejectionInfo
 from src.api.endpoints.review.enums import RejectionReason
+from src.api.endpoints.review.next.dto import GetNextURLForFinalReviewOuterResponse
+from src.api.endpoints.review.reject.dto import FinalReviewRejectionInfo
 from src.collectors.enums import URLStatus
 from src.db.models.instantiations.url.core import URL
 from tests.helpers.complex_test_data_functions import setup_for_get_next_url_for_final_review
