+++ tests/automated/integration/db/client/get_next_url_for_final_review/test_basic.py
@@ -37,7 +37,7 @@ async def test_get_next_url_for_final_review_basic(db_data_creator: DBDataCreato
 
     annotation_info = result.annotations
     relevant_info = annotation_info.relevant
-    assert relevant_info.auto == True
+    assert relevant_info.auto.is_relevant == True
     assert relevant_info.user == SuggestedStatus.NOT_RELEVANT
 
     record_type_info = annotation_info.record_type
