+++ tests/automated/integration/db/client/get_next_url_for_final_review/test_new_agency.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.api.endpoints.annotate.dtos.agency.post import URLAgencyAnnotationPostInfo
+from src.api.endpoints.annotate.agency.post.dto import URLAgencyAnnotationPostInfo
 from src.core.enums import SuggestedStatus, RecordType, SuggestionType
 from tests.helpers.batch_creation_parameters.annotation_info import AnnotationInfo
 from tests.helpers.batch_creation_parameters.core import TestBatchCreationParameters
