+++ tests/manual/external/pdap/test_sync_agencies.py
@@ -1,7 +1,7 @@
 import pytest
 import time
 
-from src.core.tasks.scheduled.operators.agency_sync.dtos import AgencySyncParameters
+from src.core.tasks.scheduled.operators.agency_sync.dtos.parameters import AgencySyncParameters
 
 
 @pytest.mark.asyncio
